+++ app/api/webhook/stripe/route.ts
@@ -1,7 +1,21 @@
+import { headers } from 'next/headers';
+import { NextResponse } from 'next/server';
 import Stripe from 'stripe';
-import { stripe } from '@/app/utils/stripe';
+import { createClient } from '@supabase/supabase-js';
 import { upsertSubscription, addCreditsToCustomer } from '@/app/utils/stripe-supabase-admin';
 import config from '@/config';
+
+if (!process.env.STRIPE_SECRET_KEY) {
+  throw new Error('Missing env.STRIPE_SECRET_KEY');
+}
+if (!process.env.STRIPE_WEBHOOK_SECRET) {
+  throw new Error('Missing env.STRIPE_WEBHOOK_SECRET');
+}
+
+const stripe = new Stripe(process.env.STRIPE_SECRET_KEY!, {
+  apiVersion: '2023-08-16',
+  typescript: true,
+});
 
 const relevantEvents = new Set([
   'customer.subscription.created',
@@ -13,97 +27,80 @@ const relevantEvents = new Set([
 
 export async function POST(req: Request) {
   const body = await req.text();
-  const sig = req.headers.get('stripe-signature') as string;
-  const webhookSecret = process.env.STRIPE_WEBHOOK_SECRET;
+  const signature = headers().get('stripe-signature');
+
   let event: Stripe.Event;
 
   try {
-    if (!sig || !webhookSecret) return;
-    event = stripe.webhooks.constructEvent(body, sig, webhookSecret);
-  } catch (err: any) {
-    console.log(`❌ Error message: ${err.message}`);
-    return new Response(`Webhook Error: ${err.message}`, { status: 400 });
+    event = stripe.webhooks.constructEvent(
+      body,
+      signature!,
+      process.env.STRIPE_WEBHOOK_SECRET!
+    );
+  } catch (err) {
+    return new NextResponse(`Webhook Error: ${err instanceof Error ? err.message : 'Unknown Error'}`, { status: 400 });
   }
 
   if (relevantEvents.has(event.type)) {
     try {
-      switch (event.type) {
-        case 'checkout.session.completed':
-          var session = event.data.object;
-          if (session.mode === 'payment' && session.payment_status === 'paid') {
-            session = await stripe.checkout.sessions.retrieve(session.id, {
-              expand: ['line_items.data.price']
-            });
-            const priceId = session.line_items.data[0].price.id;
-            const customerId = session.customer as string;
-            const plan = config.stripe.plans.find(price => price.priceId === priceId);
-            await addCreditsToCustomer(plan, customerId)
+      if (event.type === 'checkout.session.completed') {
+        const session = event.data.object as Stripe.Checkout.Session;
+        
+        if (session.mode === 'payment' && session.payment_status === 'paid') {
+          const expandedSession = await stripe.checkout.sessions.retrieve(session.id, {
+            expand: ['line_items.data.price']
+          });
+          
+          const priceId = expandedSession.line_items.data[0].price.id;
+          const customerId = expandedSession.customer as string;
+          const plan = config.stripe.plans.find(price => price.priceId === priceId);
+          await addCreditsToCustomer(plan, customerId)
+        }
+      } else if (event.type === 'customer.subscription.created' || 
+                 event.type === 'customer.subscription.updated' || 
+                 event.type === 'customer.subscription.deleted') {
+        const subscription = event.data.object as Stripe.Subscription;
+        await upsertSubscription(
+          subscription.id,
+          subscription.customer as string
+        );
+      } else if (event.type === 'invoice.payment_succeeded') {
+        const invoice = event.data.object as Stripe.Invoice;
+        
+        if (invoice.billing_reason === 'subscription_create') {
+          const subscriptionId = invoice.subscription as string;
+          const paymentIntentId = invoice.payment_intent as string;
+
+          const paymentIntent = await stripe.paymentIntents.retrieve(paymentIntentId);
+
+          try {
+            await stripe.subscriptions.update(
+              subscriptionId,
+              {
+                default_payment_method: paymentIntent.payment_method as string,
+              }
+            );
+
+            console.log("Default payment method set for subscription:", paymentIntent.payment_method);
+          } catch (err) {
+            console.error("Failed to update the default payment method for subscription:", subscriptionId, err);
           }
-          break;
-        case 'customer.subscription.created':
-        case 'customer.subscription.updated':
-        case 'customer.subscription.deleted':
-          // This event will be triggered in following scenarios:
-          // - User started a subscription
-          // - Subscribed user paid recurring fee
-          // - Subscribed user payment failed for the recurring fee
-          // - Subscribed user cancelled their subscription
-          // - Subscribed user updated their plan
-          var subscription = event.data.object;
-          await upsertSubscription(
-            subscription.id,
-            subscription.customer as string
-          );
-          break;
-        case 'invoice.payment_succeeded':
-          if (event.data.object.billing_reason == 'subscription_create') {
-            // The subscription automatically activates after successful payment
-            // Set the payment method used to pay the first invoice
-            // as the default payment method for that subscription
-            const subscriptionId = event.data.object.subscription as string;
-            const paymentIntentId = event.data.object.payment_intent as string;
-
-            // Retrieve the payment intent used to pay the subscription
-            const paymentIntent = await stripe.paymentIntents.retrieve(paymentIntentId);
-
-            try {
-              await stripe.subscriptions.update(
-                subscriptionId,
-                {
-                  default_payment_method: paymentIntent.payment_method as string,
-                },
-              );
-
-              console.log("Default payment method set for subscription:" + paymentIntent.payment_method);
-            } catch (err) {
-              console.log(err);
-              console.log(`⚠️  Falied to update the default payment method for subscription: ${subscriptionId}`);
-            }
-          }
-
-          break;
-        case 'invoice.payment_failed':
-          // If the payment fails or the customer does not have a valid payment method,
-          //  an invoice.payment_failed event is sent, the subscription becomes past_due.
-          // Use this webhook to notify your user that their payment has
-          // failed and to retrieve new card details.
-          break;
-        case 'invoice.finalized':
-          // If you want to manually send out invoices to your customers
-          // or store them locally to reference to avoid hitting Stripe rate limits.
-          break;
-        default:
-          throw new Error('Unhandled relevant event!');
+        }
+      } else if (event.type === 'invoice.payment_failed') {
+        // If the payment fails or the customer does not have a valid payment method,
+        //  an invoice.payment_failed event is sent, the subscription becomes past_due.
+        // Use this webhook to notify your user that their payment has
+        // failed and to retrieve new card details.
+      } else if (event.type === 'invoice.finalized') {
+        // If you want to manually send out invoices to your customers
+        // or store them locally to reference to avoid hitting Stripe rate limits.
+      } else {
+        throw new Error('Unhandled relevant event!');
       }
     } catch (error) {
-      console.log(error);
-      return new Response(
-        'Webhook handler failed. View your nextjs function logs.',
-        {
-          status: 400
-        }
-      );
+      console.error('Error processing webhook:', error);
+      return new NextResponse('Webhook handler failed', { status: 500 });
     }
   }
-  return new Response(JSON.stringify({ received: true }));
+  return new NextResponse(null, { status: 200 });
 }